Added poisoning effect, with potion and wand.
An "effect" is something which is added as a child to an entity, and
which then performs some kind of effect on the entity. They are immobile
and tied to their parent entity.
The example used here is of a "poisoning" effect. It will slowly remove
health from its parent.
It can be obtained either through drinking a potion of poison, or by
getting hit by a wand of poison.

+++ rules/simulation/PhysicalDomain.h
@@ -231,7 +231,7 @@ class PhysicalDomain : public Domain
 
         sigc::connection m_propertyAppliedConnection;
 
-        float m_visibilityCheckCountdown;
+        double m_visibilityCheckCountdown;
 
         BulletEntry mContainingEntityEntry;
